Patch behavior of microsoft provider to work around OIDC incompatibility of Microsoft Entra
In case you have a "custom signing key" enabled on AWS Entra, you'll need to append '?appid=APPID'
to the JWKS URL to get a list of the appropriate keys. This behavior renders Entra non-compliant
with the OIDC spec. This patch patches Ory's Microsoft Provider to work with MS Entra despite this
non-compliance with the spec.