fix: add fallback to providerLabel (#3999)
This adds a fallback to the provider label when trying to register a duplicate identifier with an oidc.

Current error message:

`Signing in will link your account to "test@test.com" at provider "". If you do not wish to link that account, please start a new login flow.`

The label represents an optional label for the UI, but in my case it's always empty. I suggest we fallback to the provider when the label is not present. In case the label is present, the behaviour won't change.

Fallback to provider:

`Signing in will link your account to "test@test.com" at provider "google". If you do not wish to link that account, please start a new login flow.`