fix: return `return_to` code if already authenticated (#4286)
This fixes a bug in native OIDC login and registration flows, where the
user already has a session in the browser the flow is continued with
(usually a web view, but depending on the platform it already has a
session cookie set). In the callback, we now correctly handle the case
in `alreadyAuthenticated` to return the session token exchange code.